[citadel] Second version
This PR proposes a new version of Citadel, which promotes
genericity and extensibility. Citadel takes care of transfering
data, it relies on a reader and writer. The former reads the data
from a given location (e.g., redis, generators), while the latter
writes the data to another location (e.g., stdout, elasticsearch).
The logic to deal with reading/writing location is included in
the corresponding readers and writers.

Examples of data transfering are available in the /bin folder.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>