[citadel] First version
This code proposes a first version of Citadel. Citadel
takes care of fetching the Perceval items from a redis
queue and storing them in Elasticsearch. Each item is
dispatched to a specific index, which is derived from
the backend name, version, category and origin of the
item. Furthermore, the index comes with an alias, which
aggregates items of the same backend, version
and category, thus allowing to query a large amount
of data to derive insightful information.

Citadel has no dependency with other GrimoireLab
components, thus promoting platform independence.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>